fix: use license key from secret
Now we store license key as a plain text inside NRIBundle CR. This
is a bad practice, because the key is visible for everyone by default.

To fix it, we can use the created secret using customSecretName and
customSecretLicenseKey options.

This change is backward compatible and shouldn't affect existing users.