chore: make licenseKey optional
Now it is mandatory to specify licenseKey in the values to start
installation immediately. But in some cases we want to postpone it,
and therefore we don't need to specify licenseKey initially.

This PR makes licenseKey parameter optional. If it's not set, we
won't create a Secret object and allow users to do it themselves later.